// src/enums/DocumentStep.ts
export enum DocumentStep {
  FACULTY_MA_UPLOAD = "FACULTY_MA_UPLOAD",
  SAR_APPROVAL = "SAR_APPROVAL",
  FACULTY_ASR_APPROVAL = "FACULTY_ASR_APPROVAL",
  DEPARTMENT_HEAD_APPROVAL = "DEPARTMENT_HEAD_APPROVAL",
  DEPARTMENT_MA_APPROVAL = "DEPARTMENT_MA_APPROVAL",
  DEPARTMENT_HEAD_RETURN = "DEPARTMENT_HEAD_RETURN",
  FACULTY_ASR_RETURN = "FACULTY_ASR_RETURN",
  SAR_RETURN = "SAR_RETURN",
  FACULTY_MA_FINAL = "FACULTY_MA_FINAL",
}